refact: client app
